FY-171/refactor: 게임 플레이 페이지 -> 퐁씬 페이지, 대진표 페이지 컴포넌트로 분리
