+++ src/pages/BracketPage.js
@@ -1,0 +1,18 @@
+import React from "react";
+import Bracket from "../components/game/Bracket";
+
+const player = (intra_id, nickname, avatar) => {
+  return {
+    intra_id: intra_id,
+    nickname: nickname,
+    avatar: avatar,
+  };
+};
+
+const BracketPage = () => {
+  const { bracketData } = useTournament();
+
+  return <Bracket subgames={bracketData.subgames} />;
+};
+
+export default BracketPage;
